+++ src/main.rs
@@ -79,7 +79,6 @@ mod mode;
 mod options;
 mod output_stream;
 mod output_target;
-mod path_ext;
 mod piece_length_picker;
 mod piece_list;
 mod platform;